requirements.txt
shinyswatch
pandas
jinja2